refactor for cleanliness
